requirements.txt
streamlit==1.35.0
pandas==2.2.2
langchain-core==0.2.10
langchain-community==0.2.5
langchain-google-genai==1.0.5
langgraph==0.0.51
python-dotenv==1.0.1
plotly==5.18.0
dash==2.16.1
dash-bootstrap-components==1.3.1
dash-daq==0.5.0
dtale==3.12.0
xarray
asteval==0.9.31
kaleido==0.2.1
openpyxl==3.1.2